also write out solution back to a new excel file

+++ reader.py
@@ -1,7 +1,5 @@
 from pathlib import Path
 import pandas as pd
 
-from openpyxl import load_workbook
-
 def read_excel(filename: Path):
-    return pd.read_excel(filename, header=None).to_numpy()
+    return pd.read_excel(filename, header=None).to_numpy().astype('float16')
